requirements.txt
discord.py
python-dotenv
google-generativeai
yt-dlp
youtube-dl
pynacl